Added watchlist and single coin page

+++ src/components/misc/WatchListButton.tsx
@@ -2,6 +2,7 @@
 
 import React from "react"
 import axios from "axios"
+import Link from "next/link"
 import { useContext } from "react"
 import { WatchlistContext } from "@/context/WatchListContext"
 import { useSession } from "next-auth/react"
@@ -16,9 +17,18 @@ type WatchListButtonProps = {
 	name?: string
 	className?: string
 	className2?: string
+	className3?: string
+	side?: "top" | "right" | "bottom" | "left"
 }
 
-export default function WatchListButton({ coinId, name, className, className2 }: WatchListButtonProps) {
+export default function WatchListButton({
+	coinId,
+	name,
+	className,
+	className2,
+	className3,
+	side,
+}: WatchListButtonProps) {
 	const [isLoading, setIsLoading] = React.useState<boolean>(false)
 	const session = useSession()
 	const {
@@ -29,6 +39,11 @@ export default function WatchListButton({ coinId, name, className, className2 }:
 	const handleWatchList = async (coinId: number) => {
 		setIsLoading(true)
 		try {
+			if (session?.status !== "authenticated") {
+				setIsLoading(false)
+				return
+			}
+
 			if (session?.data?.user && !!watchlist) {
 				if (watchlist.includes(coinId)) {
 					await axios.delete("/api/user/watchlist", {
@@ -60,19 +75,28 @@ export default function WatchListButton({ coinId, name, className, className2 }:
 	}
 
 	return (
-		<Button
-			onClick={() => handleWatchList(coinId)}
-			disabled={session.status !== "authenticated" || isLoading}
-			className={"group mr-3 p-0"}
-		>
+		<button onClick={() => handleWatchList(coinId)} className={cn("group", className3)}>
 			<Tooltip
-				side="top"
+				side={side}
 				asChild={true}
 				content={
-					watchlist?.includes(coinId) ? (
-						<p className="p-2 text-xs text-neutral-800 dark:text-neutral-100">Remove from watchlist</p>
+					session.status !== "authenticated" ? (
+						<div className="flex w-60 flex-col items-center justify-center gap-2.5 p-3">
+							<p className="text-xs font-medium text-neutral-800 dark:text-neutral-100">
+								Sign up for an account in just a few easy steps to add this coin to your watchlist!
+							</p>
+							<Link className="w-full" href={"/sign-up"}>
+								<Button className="h-8 w-full border-x border-b border-t border-emerald-400 bg-emerald-500 px-2.5 text-[.8rem] font-semibold text-white transition duration-200 ease-out hover:border-emerald-400 hover:bg-emerald-400 dark:border-emerald-500 dark:border-b-transparent dark:bg-emerald-600/80 dark:shadow-sm dark:shadow-black/30 dark:hover:border-emerald-500 dark:hover:bg-emerald-500">
+									Sign Up
+								</Button>
+							</Link>
+						</div>
+					) : watchlist?.includes(coinId) ? (
+						<p className="p-3 text-xs font-medium text-neutral-800 dark:text-neutral-100">Remove from watchlist</p>
 					) : (
-						<p className="p-2 text-xs text-neutral-800 dark:text-neutral-100">Add to watchlist and follow coin</p>
+						<p className="p-3 text-xs font-medium text-neutral-800 dark:text-neutral-100">
+							Add to watchlist and follow coin
+						</p>
 					)
 				}
 			>
@@ -80,12 +104,17 @@ export default function WatchListButton({ coinId, name, className, className2 }:
 					<Loader2 className={cn("h-3 w-3 animate-spin text-neutral-800 dark:text-neutral-100", className2)} />
 				) : (
 					<Star
-						className={cn("h-3 w-3 transition duration-200 ease-out group-hover:stroke-emerald-500", className, {
-							"fill-emerald-500 stroke-emerald-500": watchlist?.includes(coinId),
-						})}
+						className={cn(
+							"h-[13px] w-[13px] stroke-neutral-700/60 transition duration-200 ease-out group-hover:stroke-yellow-400 dark:stroke-neutral-400",
+							className,
+							{
+								"fill-yellow-400 stroke-yellow-400 dark:fill-yellow-400 dark:stroke-yellow-400":
+									watchlist.includes(coinId),
+							}
+						)}
 					/>
 				)}
 			</Tooltip>
-		</Button>
+		</button>
 	)
 }
